refactor: use `transport` for wagmi; use `strict` mode & silence type warnings
the `useSigner.ts` hook was unable to use wagmi properly due to a type issue. it was resolved with a tsconfig.json `strict` compilation, which then caused the rest of the project to be affected. thus, the types were cleaned up (lazily, by assigning `any` in most instances) such that everything works with the new wamgi upgrade

Signed-off-by: Dan Buchholz <dbuchholz30@gmail.com>